Remove PropertyGroup interface
