+++ java/configuration/src/main/java/sleeper/configuration/properties/PropertyGroup.java
@@ -16,8 +16,91 @@
 
 package sleeper.configuration.properties;
 
-public interface PropertyGroup {
-    String getName();
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Objects;
+import java.util.function.Consumer;
 
-    String getDescription();
+public class PropertyGroup {
+    private static final List<PropertyGroup> ALL = new ArrayList<>();
+    private final String name;
+    private final String description;
+
+    private PropertyGroup(Builder builder) {
+        name = Objects.requireNonNull(builder.name, "name must not be null");
+        description = Objects.requireNonNull(builder.description, "description must not be null");
+    }
+
+    public static Builder group(String name) {
+        return new Builder().name(name);
+    }
+
+    public static List<PropertyGroup> all() {
+        return Collections.unmodifiableList(ALL);
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        PropertyGroup that = (PropertyGroup) o;
+        return name.equals(that.name) && description.equals(that.description);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name, description);
+    }
+
+    @Override
+    public String toString() {
+        return description;
+    }
+
+    static final class Builder {
+        private String name;
+        private String description;
+        private Consumer<PropertyGroup> addToList = Builder::addToAllList;
+
+        private Builder() {
+        }
+
+        public Builder name(String name) {
+            this.name = name;
+            return this;
+        }
+
+        public Builder description(String description) {
+            this.description = description;
+            return this;
+        }
+
+        public Builder addToList(Consumer<PropertyGroup> addToList) {
+            this.addToList = addToList;
+            return this;
+        }
+
+        public PropertyGroup build() {
+            PropertyGroup group = new PropertyGroup(this);
+            addToList.accept(group);
+            return group;
+        }
+
+        private static void addToAllList(PropertyGroup property) {
+            ALL.add(property);
+        }
+    }
 }
